move dma_display to hal
