+++ gateware/src/example_soc/fw/src/lib.rs
@@ -16,5 +16,5 @@ tiliqua_hal::impl_i2c! {
     I2c0: pac::I2C0,
 }
 
-pub mod log;
+pub mod handlers;
 pub mod opts;
