+++ gateware/src/rs/hal/Cargo.toml
@@ -22,3 +22,4 @@ embedded-hal-nb = "=1.0.0"
 log = { version = "=0.4.17", optional = true }
 nb = "=1.1.0"
 riscv = { version = "=0.10.1" }
+embedded-graphics = "0.8.1"
