+++ gateware/src/rs/hal/src/lib.rs
@@ -8,6 +8,7 @@
 pub mod serial;
 pub mod timer;
 pub mod i2c;
+pub mod dma_display;
 
 pub use embedded_hal as hal;
 pub use embedded_hal_nb as hal_nb;
