+++ gateware/src/rs/lib/src/lib.rs
@@ -2,3 +2,4 @@
 
 pub mod opt;
 pub mod draw;
+pub mod log;
